simplify and clean code

+++ core/src/main/scala/io/qbeast/core/model/NormalizedWeight.scala
@@ -20,7 +20,7 @@ object NormalizedWeight {
    * @param cubeSize the actual cube size, must not be zero
    * @return a normalized weight
    */
-  def apply(desiredCubeSize: Double, cubeSize: Long): NormalizedWeight = {
+  def apply(desiredCubeSize: Int, cubeSize: Long): NormalizedWeight = {
     if (cubeSize == 0) {
       throw new IllegalArgumentException("Cube size is zero.")
     }
